+++ exchangebatch_test.go
@@ -25,7 +25,7 @@ var _ = Describe("func Exchange() (batch requests)", func() {
 		reader                       *RequestSetReaderStub
 		writer                       *ResponseWriterStub
 		logs                         *observer.ObservedLogs
-		logger                       *ZapExchangeLogger
+		logger                       ExchangeLogger
 		closed                       bool
 	)
 
@@ -85,10 +85,7 @@ var _ = Describe("func Exchange() (batch requests)", func() {
 
 		var core zapcore.Core
 		core, logs = observer.New(zapcore.DebugLevel)
-
-		logger = &ZapExchangeLogger{
-			Target: zap.New(core),
-		}
+		logger = NewZapExchangeLogger(zap.New(core))
 
 		closed = false
 	})