+++ exchangesingle_test.go
@@ -21,7 +21,7 @@ var _ = Describe("func Exchange() (single request)", func() {
 		reader    *RequestSetReaderStub
 		writer    *ResponseWriterStub
 		logs      *observer.ObservedLogs
-		logger    *ZapExchangeLogger
+		logger    ExchangeLogger
 		closed    bool
 	)
 
@@ -74,10 +74,7 @@ var _ = Describe("func Exchange() (single request)", func() {
 
 		var core zapcore.Core
 		core, logs = observer.New(zapcore.DebugLevel)
-
-		logger = &ZapExchangeLogger{
-			Target: zap.New(core),
-		}
+		logger = NewZapExchangeLogger(zap.New(core))
 
 		closed = false
 	})