+++ logging.go
@@ -8,6 +8,7 @@ import (
 
 	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
+	"golang.org/x/exp/slog"
 )
 
 // ExchangeLogger is an interface for logging JSON-RPC requests, responses and
@@ -28,118 +29,145 @@ type ExchangeLogger interface {
 	LogCall(ctx context.Context, req Request, res Response)
 }
 
-// ZapExchangeLogger is an implementation of ExchangeLogger using zap.Logger.
-type ZapExchangeLogger struct {
-	// Target is the destination for log messages.
-	Target *zap.Logger
+// NewZapExchangeLogger returns an ExchangeLogger that targets the given
+// [zap.Logger].
+func NewZapExchangeLogger(t *zap.Logger) ExchangeLogger {
+	return &structuredExchangeLogger[zap.Field]{
+		Target: t,
+		Int:    zap.Int,
+		String: zap.String,
+	}
 }
 
-var _ ExchangeLogger = (*ZapExchangeLogger)(nil)
+// NewSLogExchangeLogger returns an ExchangeLogger that targets the given
+// [slog.Logger].
+func NewSLogExchangeLogger(t *slog.Logger) ExchangeLogger {
+	return &structuredExchangeLogger[any]{
+		Target: t,
+		Int: func(n string, v int) any {
+			return slog.Int(n, v)
+		},
+		String: func(n string, v string) any {
+			return slog.String(n, v)
+		},
+	}
+}
+
+type structuredExchangeLogger[Attr any] struct {
+	Target interface {
+		Info(message string, attrs ...Attr)
+		Error(message string, attrs ...Attr)
+	}
+	Int    func(string, int) Attr
+	String func(string, string) Attr
+}
+
+var _ ExchangeLogger = (*structuredExchangeLogger[any])(nil)
 
 // LogError writes an information about an error response that is a result of
 // some problem with the request set as a whole.
-func (l ZapExchangeLogger) LogError(ctx context.Context, res ErrorResponse) {
-	fields := []zap.Field{
-		zap.Int("error_code", int(res.Error.Code)),
-		zap.String("error", res.Error.Code.String()),
+func (l structuredExchangeLogger[Attr]) LogError(ctx context.Context, res ErrorResponse) {
+	attrs := []Attr{
+		l.Int("error_code", int(res.Error.Code)),
+		l.String("error", res.Error.Code.String()),
 	}
 
 	if span := trace.SpanFromContext(ctx); span.IsRecording() {
-		fields = append(fields, zap.String("trace_id", span.SpanContext().TraceID().String()))
+		attrs = append(attrs, l.String("trace_id", span.SpanContext().TraceID().String()))
 	}
 
 	if res.ServerError != nil {
-		fields = append(fields, zap.String("caused_by", res.ServerError.Error()))
+		attrs = append(attrs, l.String("caused_by", res.ServerError.Error()))
 	}
 
 	if res.Error.Message != res.Error.Code.String() {
-		fields = append(fields, zap.String("responded_with", res.Error.Message))
+		attrs = append(attrs, l.String("responded_with", res.Error.Message))
 	}
 
 	l.Target.Error(
 		"error",
-		fields...,
+		attrs...,
 	)
 }
 
 // LogWriterError logs about an error that occured when attempting to use a
 // ResponseWriter.
-func (l ZapExchangeLogger) LogWriterError(ctx context.Context, err error) {
-	fields := []zap.Field{
-		zap.String("error", err.Error()),
+func (l structuredExchangeLogger[Attr]) LogWriterError(ctx context.Context, err error) {
+	attrs := []Attr{
+		l.String("error", err.Error()),
 	}
 
 	if span := trace.SpanFromContext(ctx); span.IsRecording() {
-		fields = append(fields, zap.String("trace_id", span.SpanContext().TraceID().String()))
+		attrs = append(attrs, l.String("trace_id", span.SpanContext().TraceID().String()))
 	}
 
 	l.Target.Error(
 		"unable to write JSON-RPC response",
-		fields...,
+		attrs...,
 	)
 }
 
 // LogNotification logs information about a notification request.
-func (l ZapExchangeLogger) LogNotification(ctx context.Context, req Request) {
+func (l structuredExchangeLogger[Attr]) LogNotification(ctx context.Context, req Request) {
 	var w strings.Builder
 
 	w.WriteString("notify ")
 	writeMethod(&w, req.Method)
 
-	fields := []zap.Field{
-		zap.Int("param_size", len(req.Parameters)),
+	attrs := []Attr{
+		l.Int("param_size", len(req.Parameters)),
 	}
 
 	if span := trace.SpanFromContext(ctx); span.IsRecording() {
-		fields = append(fields, zap.String("trace_id", span.SpanContext().TraceID().String()))
+		attrs = append(attrs, l.String("trace_id", span.SpanContext().TraceID().String()))
 	}
 
 	l.Target.Info(
 		w.String(),
-		fields...,
+		attrs...,
 	)
 }
 
 // LogCall logs information about a call request and its response.
-func (l ZapExchangeLogger) LogCall(ctx context.Context, req Request, res Response) {
+func (l structuredExchangeLogger[Attr]) LogCall(ctx context.Context, req Request, res Response) {
 	var w strings.Builder
 
 	w.WriteString("call ")
 	writeMethod(&w, req.Method)
 
-	fields := []zap.Field{
-		zap.Int("param_size", len(req.Parameters)),
+	attrs := []Attr{
+		l.Int("param_size", len(req.Parameters)),
 	}
 
 	if span := trace.SpanFromContext(ctx); span.IsRecording() {
-		fields = append(fields, zap.String("trace_id", span.SpanContext().TraceID().String()))
+		attrs = append(attrs, l.String("trace_id", span.SpanContext().TraceID().String()))
 	}
 
 	switch res := res.(type) {
 	case SuccessResponse:
-		fields = append(fields, zap.Int("result_size", len(res.Result)))
+		attrs = append(attrs, l.Int("result_size", len(res.Result)))
 		l.Target.Info(
 			w.String(),
-			fields...,
+			attrs...,
 		)
 	case ErrorResponse:
-		fields = append(
-			fields,
-			zap.Int("error_code", int(res.Error.Code)),
-			zap.String("error", res.Error.Code.String()),
+		attrs = append(
+			attrs,
+			l.Int("error_code", int(res.Error.Code)),
+			l.String("error", res.Error.Code.String()),
 		)
 
 		if res.ServerError != nil {
-			fields = append(fields, zap.String("caused_by", res.ServerError.Error()))
+			attrs = append(attrs, l.String("caused_by", res.ServerError.Error()))
 		}
 
 		if res.Error.Message != res.Error.Code.String() {
-			fields = append(fields, zap.String("responded_with", res.Error.Message))
+			attrs = append(attrs, l.String("responded_with", res.Error.Message))
 		}
 
 		l.Target.Error(
 			w.String(),
-			fields...,
+			attrs...,
 		)
 	}
 }
@@ -184,31 +212,4 @@ func writeDataSize(w *strings.Builder, n int) {
 	}
 
 	fmt.Fprintf(w, "%0.1f PB", f/1000)
-}
-
-// writeRequestDetails writes the details of a request to w.
-func writeRequestDetails(w *strings.Builder, req Request) {
-	w.WriteString("params: ")
-	writeDataSize(w, len(req.Parameters))
-}
-
-// writeSuccessResponseDetails writes the details of a success response to w.
-func writeSuccessResponseDetails(w *strings.Builder, res SuccessResponse) {
-	w.WriteString("result: ")
-	writeDataSize(w, len(res.Result))
-}
-
-// writeErrorResponseDetails writes the details of an error response to w.
-func writeErrorResponseDetails(w *strings.Builder, res ErrorResponse) {
-	fmt.Fprintf(w, "error: %d %s", res.Error.Code, res.Error.Code.String())
-
-	if res.ServerError != nil {
-		w.WriteString(", caused by: ")
-		w.WriteString(res.ServerError.Error())
-	}
-
-	if res.Error.Message != res.Error.Code.String() {
-		w.WriteString(", responded with: ")
-		w.WriteString(res.Error.Message)
-	}
 }