+++ CHANGELOG.md
@@ -10,6 +10,17 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## Unreleased
+
+### Added
+
+- Added `NewZapExchangeLogger()`
+- Added `NewSLogExchangeLogger()`
+
+### Removed
+
+- Removed `ZapExchangeLogger` struct
+
 ## [0.8.2] - 2023-04-26
 
 ### Changed